fix: Update package version, move component to new file, update home page style
