+++ package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@clerk/clerk-react": "^4.32.2",
-    "@clerk/nextjs": "^4.31.4",
+    "@clerk/nextjs": "^4.31.3",
     "@headlessui/react": "^1.7.19",
     "@hookform/resolvers": "^3.6.0",
     "@libsql/client": "0.4.0-pre.7",
@@ -32,7 +32,7 @@
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.51.5",
     "react-icons": "^4.12.0",
-    "redis": "^4.7.0",
+    "redis": "^4.6.14",
     "sharp": "^0.33.4",
     "square": "^34.0.1",
     "swr": "^2.2.5",
@@ -63,7 +63,7 @@
     "simple-git-hooks": "^2.11.1",
     "tailwindcss": "^3.4.4",
     "tsec": "^0.2.8",
-    "typescript": "^5.3.0"
+    "typescript": "^5.4.5"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx,css,json,mjs,cjs,yml,yaml}": [
