+++ src/app/(home)/page.tsx
@@ -8,39 +8,7 @@ import { SPONSOR_TYPES, getSponsors } from '@/data/sponsors';
 import Image from 'next/image';
 import Link from 'next/link';
 import { Fragment } from 'react';
-import { FiClock, FiMapPin } from 'react-icons/fi';
-
-function UpcomingEventCard({ event, index }: { event: Event; index: number }) {
-    return (
-        <FancyRectangle colour="white" offset="8" rounded fullWidth>
-            <div className="flex w-full flex-col gap-6 rounded-xl bg-white p-4 text-black lg:flex-row">
-                <div className="grow space-y-2 md:space-y-4">
-                    <div className="flex gap-6">
-                        <div
-                            className={`h-fit rounded-md border-[3px] border-black px-4 py-2 text-xl ${['bg-orange', 'bg-yellow', 'bg-purple'][index % 3]}`}
-                        >
-                            <div>{event.date.month}</div>
-                            <div>{event.date.day}</div>
-                        </div>
-                        <div className="grow space-y-2 text-2xl">
-                            <h4 className=" md:border-b-[3px] md:border-black md:pb-1 ">
-                                {event.title}
-                            </h4>
-                            <div className="flex gap-2">
-                                <FiClock size={26} />
-                                <span>{event.time}</span>
-                            </div>
-                            <div className="flex gap-2">
-                                <FiMapPin size={26} />
-                                <span>{event.location}</span>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </FancyRectangle>
-    );
-}
+import UpcomingEventCard from './UpcomingEventCard';
 
 const getEventDate = (event: Event) => {
     return new Date(
@@ -110,9 +78,9 @@ export default function HomePage() {
                         className={`flex flex-col justify-around ${UPCOMING_EVENTS.length === 0 ? 'lg:col-span-2' : ''}`}
                     >
                         <div>
-                            <div className="relative z-10 mt-12 flex flex-col text-2xl font-black md:flex-row lg:mt-24 lg:text-3xl">
+                            <div className="relative z-10 mt-12 flex flex-col text-2xl font-black lg:mt-24 lg:text-3xl">
                                 <h3>New Members are</h3>
-                                <div className="mt-2 w-fit bg-purple px-2 md:ml-2 md:mt-0">
+                                <div className="mt-2 w-fit bg-purple px-2 ">
                                     <h3 className=" text-grey">Always Welcome</h3>
                                 </div>
                             </div>
@@ -147,10 +115,9 @@ export default function HomePage() {
                         </div>
                     </div>
 
-                    {/* <div className="relative z-10 mt-12 text-2xl font-black md:flex-row lg:ml-4 lg:mt-24 lg:text-3xl"> */}
                     {UPCOMING_EVENTS.length > 0 && (
-                        <div className="relative z-10 mt-12 text-2xl font-black md:flex-row lg:ml-4 lg:mt-24 lg:text-3xl">
-                            <div className="flex justify-end">
+                        <div className="relative z-10 mt-12 text-2xl font-black md:flex-row lg:ml-10 lg:mt-24 lg:text-3xl">
+                            <div className="flex md:justify-end">
                                 <h3>Upcoming Events</h3>
                             </div>
 
@@ -161,7 +128,6 @@ export default function HomePage() {
                             </div>
                         </div>
                     )}
-                    {/* </div> */}
                 </div>
             </section>
